feat: GoogleServicesHelper
Created a google play services helper class to encapsulate the code that handles connecting with google play services.  Updated the Adapter and MainActivity to use the Helper class to function differently according to whether the application is connected to Google play services.  The Activity also connects and disconnects the service during onStart & onStop lifecycle methods.

Stage 3 of Treehouse tutorial